requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0
pandas>=2.0.0
supabase>=2.22.0
python-dotenv>=1.0.0
psutil>=7.0.0